+++ src/block-types/row/save.js
@@ -1,9 +1,17 @@
 import { useBlockProps, InnerBlocks } from "@wordpress/block-editor";
+import spacing from "../../helpers/spacing";
 import flexbox from "../../helpers/flexbox";
 
 const save = (props) => {
+	const {
+		attributes: { removeGutters },
+	} = props;
+
 	const blockProps = useBlockProps.save({
-		className: `row${flexbox.cssClasses(props)}`,
+		className: `row${removeGutters ? " no-gutters" : ""}${flexbox.cssClasses(
+			props
+		)}${spacing.cssClasses(props)}`,
+		style: spacing.cssVars(props, "row"),
 	});
 
 	return (